requirements.txt
gradio==5.49.1
langchain_core==1.0.1
langchain_google_genai==3.0.0
langchain_tavily==0.2.12
langchain_text_splitters==1.0.0
langgraph==1.0.1
pinecone==7.3.0
pypdf==6.1.3
python-dotenv==1.2.1
sentence_transformers==5.0.0
protobuf==3.20.1